L6: first pass at phi pass
My initial plan to formulate the SSA transformation as standard compiler
passes was not long for this world. Both the phi-adding and the
variable-renaming want to traverse the CFG in a specific order.

So add the skeleton of a proper SSA transformation, implement the
phi/upsilon pass, and start implementing the variable renaming pass.