PROJECT: move decorator experiment to more appropriately named file
The decorator experiment was tough because compiling functions in
isolation seems to be non-trivial. In general, it isn't even clear what
that would mean, given the semantics of Python (how should free
variables in functions be bound, for example). Googling around, people
seem to solve this problem in rewrite-decorators by re-compiling the
enclosing file and plucking out the code for the decorated function and
sticking that back into the larger module for the original file. That
seems like more work than I want to do.

So I am moving on to a file-rewriting model! Which should be much
simpler because
1) I don't have to compile functions in isolation, and
2) I might not have to compile anything! I can probably just write out a
   new file!