PROJECT: update test file
Add more test cases. At this point I am "testing" by running the pass on
the file and visually inspecting that it is making the expected
transformations, so "test coverage" is measured in unique AST patterns tested,
not anything like runtime of the test functions. This is all to say that
test coverage has been increased by adding some new more-or-less useless
functions to the test file.

Also commit the current result of running the pass on the test file.